Refactor with decorator
