+++ app/decorators/audit_activity/test/result_decorator.rb
@@ -1,0 +1,12 @@
+class AuditActivity::Test::ResultDecorator < ApplicationDecorator
+  delegate_all
+
+  def title(_viewing_user = nil)
+    return super if metadata.nil?
+
+    # The decorator provides logic to create the title but it may have changed
+    # in subsequent edits, so we need to create a new instance using the
+    # original attributes in order to decorate it
+    Test::Result.new(metadata["test_result"].except("document")).decorate.title
+  end
+end
