+++ app/models/audit_activity/test/result.rb
@@ -9,15 +9,6 @@ class AuditActivity::Test::Result < AuditActivity::Test::Base
     raise "Deprecated - use AddTestResultToInvestigation.call instead"
   end
 
-  def title(_viewing_user = nil)
-    return super if metadata.nil?
-
-    # The decorator provides logic to create the title but it may have changed
-    # in subsequent edits, so we need to create a new instance using the
-    # original attributes in order to decorate it
-    Test::Result.new(metadata["test_result"].except("document")).decorate.title
-  end
-
   def test_result
     return if metadata.nil?
